+++ src/GRA.Domain.Model/AvatarItemText.cs
@@ -6,7 +6,15 @@ namespace GRA.Domain.Model
     {
         [MaxLength(130)]
         public string AltText { get; set; }
+
         public int AvatarItemId { get; set; }
+
         public int LanguageId { get; set; }
+
+        /// <summary>
+        /// This variable does not map back to the database and is only used to deserialize items
+        /// during avatar import.
+        /// </summary>
+        public string LanguageName { get; set; }
     }
 }