+++ src/GRA.Domain.Model/Utility/AvatarBundleTransfer.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace GRA.Domain.Model.Utility
+{
+    public class AvatarBundleTransfer
+    {
+        public ICollection<AvatarBundleItemTransfer> AvatarItems { get; set; }
+        public string Name { get; set; }
+    }
+}