+++ src/GRA.Domain.Service/Abstract/BaseService.cs
@@ -1,21 +1,20 @@
 ﻿using System;
+using System.Threading.Tasks;
 using GRA.Abstract;
+using GRA.Domain.Model;
+using GRA.Domain.Repository;
 using Microsoft.Extensions.Logging;
 
 namespace GRA.Domain.Service.Abstract
 {
-    public abstract class BaseService<TService>
+    public abstract class BaseService<TService>(ILogger<TService> logger,
+        IDateTimeProvider dateTimeProvider)
     {
-        protected readonly IDateTimeProvider _dateTimeProvider;
-        protected readonly ILogger<TService> _logger;
+        protected readonly IDateTimeProvider _dateTimeProvider = dateTimeProvider
+            ?? throw new ArgumentNullException(nameof(dateTimeProvider));
 
-        protected BaseService(ILogger<TService> logger,
-            IDateTimeProvider dateTimeProvider)
-        {
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-            _dateTimeProvider = dateTimeProvider
-                ?? throw new ArgumentNullException(nameof(dateTimeProvider));
-        }
+        protected readonly ILogger<TService> _logger = logger
+            ?? throw new ArgumentNullException(nameof(logger));
 
         protected static string GetCacheKey(string cacheKey, params object[] cacheKeyValues)
         {
@@ -27,6 +26,18 @@ namespace GRA.Domain.Service.Abstract
         protected static int GetPercent(int count, int total)
         {
             return count * 100 / total;
+        }
+
+        protected static async Task ReportJobStatusAsync(IJobRepository repo,
+            JobMetadata metadata,
+            JobStatus status)
+        {
+            ArgumentNullException.ThrowIfNull(repo);
+            ArgumentNullException.ThrowIfNull(metadata);
+            ArgumentNullException.ThrowIfNull(status);
+
+            await repo.UpdateStatusAsync(metadata.JobId, status.Status);
+            metadata.Progress.Report(status);
         }
     }
 }